requirements.txt
asgiref==3.6.0
Babel==2.12.1
black==23.1.0
click==8.1.3
Django==4.1.6
django-appconf==1.0.5
django-phone-field==1.8.1
django-widget-tweaks==1.4.12
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.11
Pillow==9.4.0
platformdirs==3.0.0
pytz==2022.7.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.3
tomli==2.0.1


